frontend/admin-dashboard/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
bcrypt==4.1.2
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
